soal_1/test1/requirements.txt
faker
djangorestframework
grpcio
grpcio-tools
uvicorn